requirements.txt
discord
colorama
art
python_dotenv
